backend/requirements.txt
flask
flask-cors
torch
torchvision
diffusers
transformers
timm
opencv-python
numpy
Pillow